pyproject.toml
[tool.poetry]
name = "scriptmaker"
version = "1.0.0"
description = "A Python library for producing custom scripts in Blood On The Clocktower."
authors = ["Rohan Sarvaria <rsarvaria@uwaterloo.ca>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "scriptmaker" }]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.1.0"
jinja2 = "^3.1.2"
weasyprint = "^60.1"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.335"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
